test: split test_sdlc_session_ensure into per-theme modules (Refs #2879)
