+++ tests/unit/sdlc_session_ensure/__init__.py
@@ -1,0 +1,1 @@
+"""Unit tests for tools.sdlc_session_ensure, split by theme (#2879)."""
